refactor: split types into shared/eslint/oxlint and port comments config
- Create eslint/types.ts and oxlint/types.ts, trim shared types.ts
- Move OptionsFormatters to shared types
- Port ESLint comments config to oxlint 1:1 (two config items)
