+++ src/eslint/configs/jsx.ts
@@ -1,5 +1,5 @@
 import { GLOB_JSX, GLOB_TSX } from "../../globs";
-import type { TypedFlatConfigItem } from "../../types";
+import type { TypedFlatConfigItem } from "../types";
 
 export async function jsx(): Promise<Array<TypedFlatConfigItem>> {
 	return [
