+++ src/eslint/index.ts
@@ -1,9 +1,9 @@
 export * from "../globs";
-export * from "../types";
 export { isInEditorEnvironment } from "../utils";
 export { isInAgentSession } from "../utils";
 export { isInGitHooksOrLintStaged } from "../utils";
-
 export * from "./configs";
+
 export * from "./factory";
 export { isentinel as default } from "./factory";
+export * from "./types";
