+++ src/oxlint/configs/javascript.ts
@@ -1,7 +1,7 @@
 import globals from "globals";
 
 import { GLOB_SRC } from "../../globs.ts";
-import type { Rules, TypedOxlintConfigItem } from "../../types.ts";
+import type { Rules, TypedOxlintConfigItem } from "../types.ts";
 
 export function oxlintJavascript(options: {
 	isInEditor: boolean;
