+++ src/oxlint/configs/promise.ts
@@ -1,5 +1,5 @@
 import { GLOB_SRC } from "../../globs.ts";
-import type { Rules, TypedOxlintConfigItem } from "../../types.ts";
+import type { Rules, TypedOxlintConfigItem } from "../types.ts";
 
 export function oxlintPromise(): Array<TypedOxlintConfigItem> {
 	// Inlined from promiseRules()
